+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Allow the user to select which page to debug in case multiple are available ([#2][i2])
+
 ## 1.1.1 - 2019-09-04
 * Change the `extensionKind` to `ui` because vscode-chrome-debug can not be installed on a remote system
 * Add a check for VS LiveShare to prevent the debugging configuration from being resolved in a guest instance